I moved load_action function for loader py file
